[feat] 기획에 맞게 UII 수정, 로직분리
스크립트별로 로직 분리 (리펙토링)
기획에 맞춰 UI 변경.
메인씬 펫 유전자정보 UI 연결

+++ Assets/Scripts/GameSystem/SaveSystem.cs
@@ -179,7 +179,7 @@ public class UserData
     public Language CurLanguage;
     public string UserDisplayName;
     public int MaxPetAmount;
-    public float Energy;
+    public int Energy;
     public List<EggData> EggList;
     public List<PetSaveData> HavePetList;
     public List<PetRecordData> HadPetList;
